// todo-webapp/src/generated/todo-api.ts
export interface components {
    schemas: {
        Todo: {
            id: string;
            description: string;
            done: boolean;
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            completedAt?: string | null;
        };
        NewTodo: {
            description: string;
        };
        Error: {
            /** @description HTTP or application error code */
            code: number;
            /** @description short human-readable label */
            message: string;
            /** @description detailed explanation */
            description?: string;
            /** @description URI to documentation */
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}